+++ packages/util-crypto/src/secp256k1/compress.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2023 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { Point } from '@noble/secp256k1';
+import { secp256k1 } from '@noble/curves/secp256k1';
 
 import { hasBigInt } from '@polkadot/util';
 import { isReady, secp256k1Compress as wasm } from '@polkadot/wasm-crypto';
@@ -17,5 +17,5 @@ export function secp256k1Compress (publicKey: Uint8Array, onlyJs?: boolean): Uin
 
   return !hasBigInt || (!onlyJs && isReady())
     ? wasm(publicKey)
-    : Point.fromHex(publicKey).toRawBytes(true);
+    : secp256k1.ProjectivePoint.fromHex(publicKey).toRawBytes(true);
 }